+++ r-naming.tf
@@ -31,3 +31,15 @@ resource "azurecaf_name" "servicebus_namespace_auth_rule" {
   clean_input   = true
   separator     = "-"
 }
+
+resource "azurecaf_name" "servicebus_topic" {
+  for_each = { for t in var.servicebus_topics : t.name => t }
+
+  name          = var.stack
+  resource_type = "azurerm_servicebus_topic"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, each.key, local.name_suffix])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}